refactor(cli): extract default run flow to src/cli/commands/run.mjs
main() の default run 分岐（skill-set 解決・planLocalReview・runLocalReview・
save/digest/baseline・出力整形・gate exit）を runRunCommand(parsed, targetPath) へ
抽出（#1592 と同型）。dispatch は return await で main() の outer try/catch を経由させ、
非 git ディレクトリでの friendly エラー（Not a git repository + Hints, exit 1）を維持する。
動的 import の相対深度のみ調整し、振る舞い・出力・exit code は不変。
main() は 411→122 行に縮小。
